% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{clean_variable_names}
\alias{clean_variable_names}
\title{Clean Variable Names}
\usage{
clean_variable_names(names)
}
\arguments{
\item{names}{Character vector of variable names}
}
\value{
Character vector of cleaned names
}
\description{
Cleans and standardizes variable names for consistency.
}
\examples{
clean_variable_names(c("var.1", "var..2", "var...3"))
}
