% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{create_dimension_mapping_optimized}
\alias{create_dimension_mapping_optimized}
\title{Create optimized dimension mapping}
\usage{
create_dimension_mapping_optimized(dataset_dimensions)
}
\description{
Create optimized dimension mapping
}
\keyword{internal}
