% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{determine_frequency}
\alias{determine_frequency}
\title{Determine Data Frequency}
\usage{
determine_frequency(dates)
}
\arguments{
\item{dates}{A vector of dates (Date, POSIXct, or coercible to Date)}
}
\value{
Numeric frequency value: 12 for monthly, 4 for quarterly, 1 for annual/other
}
\description{
Determines the frequency of time series data based on typical intervals
between observations. This is used internally for time series analysis
and forecasting functions.
}
\keyword{internal}
