% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\docType{data}
\name{.metadata_cache}
\alias{.metadata_cache}
\title{In-memory metadata cache environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.metadata_cache
}
\description{
In-memory metadata cache environment
}
\keyword{internal}
