% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_multiple_datasets}
\alias{download_multiple_datasets}
\title{Download Multiple Datasets}
\usage{
download_multiple_datasets(
  dataset_ids,
  filter = "ALL",
  start_time = "",
  n_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{dataset_ids}{Character vector of ISTAT dataset IDs}

\item{filter}{Character string specifying data filters. Default is "ALL"}

\item{start_time}{Character string specifying the start period}

\item{n_cores}{Integer number of cores to use for parallel processing.
Default is parallel::detectCores() - 1}
}
\value{
A named list of data.tables, one for each dataset
}
\description{
Downloads multiple datasets from ISTAT SDMX API in parallel.
}
\examples{
\dontrun{
# Download multiple datasets
datasets <- c("534_50", "534_51", "534_52")
data_list <- download_multiple_datasets(datasets, start_time = "2020")
}
}
