% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{download_single_codelist}
\alias{download_single_codelist}
\title{Download Single Codelist}
\usage{
download_single_codelist(dataset_id)
}
\arguments{
\item{dataset_id}{Character string specifying the dataset ID}
}
\value{
A data.table containing the codelist
}
\description{
Internal function to download codelist for a single dataset.
}
\keyword{internal}
