% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_root_id}
\alias{extract_root_id}
\title{Extract Root Dataset ID}
\usage{
extract_root_id(dataset_id)
}
\arguments{
\item{dataset_id}{Character string, full dataset ID}
}
\value{
Character string, root dataset ID
}
\description{
Extracts the root dataset ID from compound IDs.
For example: "534_49_DF_DCSC_GI_ORE_10" -> "534_49"
}
\keyword{internal}
