% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{fetch_dataflow_endpoint}
\alias{fetch_dataflow_endpoint}
\title{Download Dataflows from ISTAT Dataflow Endpoint}
\usage{
fetch_dataflow_endpoint(force_update = FALSE, cache_dir = NULL)
}
\arguments{
\item{force_update}{Logical to force metadata update}

\item{cache_dir}{Character string for cache directory}
}
\value{
A data.table containing dataflow information
}
\description{
Downloads dataset metadata using the ISTAT SDMX dataflow endpoint:
https://esploradati.istat.it/SDMXWS/rest/dataflow
}
\examples{
\dontrun{
# Get all available dataflows
dataflows <- fetch_dataflow_endpoint()

# Filter to find employment-related datasets
employment_flows <- dataflows[grepl("employment|lavoro", name, ignore.case = TRUE)]
}
}
