% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{fetch_multiple_data_endpoint}
\alias{fetch_multiple_data_endpoint}
\title{Download Multiple Datasets from Data Endpoint}
\usage{
fetch_multiple_data_endpoint(dataset_ids, ...)
}
\arguments{
\item{dataset_ids}{Character vector of dataset IDs}

\item{...}{Additional parameters passed to fetch_data_endpoint}
}
\value{
A named list of data.tables
}
\description{
Downloads multiple datasets in parallel using the data endpoint.
}
\examples{
\dontrun{
# Download multiple datasets by category
employment_ids <- get_dataset_category("employment")
employment_data <- fetch_multiple_data_endpoint(employment_ids)
}
}
