% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{fetch_registry_dimensions}
\alias{fetch_registry_dimensions}
\title{Get Dataset Dimensions from Registry Endpoint}
\usage{
fetch_registry_dimensions(dataset_id)
}
\arguments{
\item{dataset_id}{Character string specifying dataset ID}
}
\value{
A list of dataset dimensions
}
\description{
Retrieves dataset dimensions using the ISTAT SDMX registry capabilities.
}
\examples{
\dontrun{
# Get dimensions for employment dataset
dims <- fetch_registry_dimensions("150_908")
}
}
