% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_by_time}
\alias{filter_by_time}
\title{Filter Data by Time Period}
\usage{
filter_by_time(data, start_date = NULL, end_date = NULL, time_col = "tempo")
}
\arguments{
\item{data}{A data.table with time information}

\item{start_date}{Date or character string specifying start date}

\item{end_date}{Date or character string specifying end date}

\item{time_col}{Character string specifying the time column name.
Default is "tempo"}
}
\value{
Filtered data.table
}
\description{
Filters ISTAT data by a specified time period.
}
\examples{
\dontrun{
# Filter data from 2020 onwards
filtered_data <- filter_by_time(data, start_date = "2020-01-01")
}
}
