% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_naive}
\alias{forecast_naive}
\title{Naive Forecast}
\usage{
forecast_naive(ts_data, periods, confidence_levels)
}
\arguments{
\item{ts_data}{A time series object}

\item{periods}{Integer number of periods to forecast}

\item{confidence_levels}{Numeric vector of confidence levels}
}
\value{
A forecast-like object
}
\description{
Performs naive forecasting (last value carried forward).
}
\keyword{internal}
