% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{generate_forecast_dates}
\alias{generate_forecast_dates}
\title{Generate Forecast Dates}
\usage{
generate_forecast_dates(historical_dates, periods, frequency)
}
\arguments{
\item{historical_dates}{Vector of historical dates}

\item{periods}{Integer number of forecast periods}

\item{frequency}{Numeric frequency of the data}
}
\value{
Vector of forecast dates
}
\description{
Generates future dates for forecasting periods.
}
\keyword{internal}
