% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_istat_provider}
\alias{get_istat_provider}
\title{Get ISTAT Service Provider}
\usage{
get_istat_provider()
}
\value{
An rsdmx SDMXServiceProvider object
}
\description{
Creates and returns an ISTAT SDMX service provider for use with rsdmx package.
}
\examples{
\dontrun{
# Get service provider
provider <- get_istat_provider()
rsdmx::addSDMXServiceProvider(provider)
}
}
