% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{integrate_downloaded_data}
\alias{integrate_downloaded_data}
\title{Integrate Downloaded Data with Existing Data}
\usage{
integrate_downloaded_data(existing_data, new_data)
}
\arguments{
\item{existing_data}{data.table of previously downloaded data}

\item{new_data}{data.table of newly downloaded data}
}
\value{
data.table with merged and deduplicated rows
}
\description{
Merges newly downloaded data with existing data, deduplicating on all
dimension columns plus ObsDimension.
}
\keyword{internal}
