% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{list_istat_endpoints}
\alias{list_istat_endpoints}
\title{List Available ISTAT Endpoints}
\usage{
list_istat_endpoints()
}
\value{
A data.frame describing available endpoints
}
\description{
Returns information about all available ISTAT SDMX endpoints.
}
\examples{
\dontrun{
# See all available endpoints
endpoints <- list_istat_endpoints()
print(endpoints)
}
}
