% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{plot.istat_forecast}
\alias{plot.istat_forecast}
\title{Plot method for istat_forecast objects}
\usage{
\method{plot}{istat_forecast}(x, include_ensemble = TRUE, include_intervals = TRUE, ...)
}
\arguments{
\item{x}{An istat_forecast object}

\item{include_ensemble}{Include ensemble forecast in plot (default TRUE)}

\item{include_intervals}{Include prediction intervals (default TRUE)}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot method for istat_forecast objects
}
