% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{print.istat_forecast}
\alias{print.istat_forecast}
\title{Print method for istat_forecast objects}
\usage{
\method{print}{istat_forecast}(x, ...)
}
\arguments{
\item{x}{An istat_forecast object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for istat_forecast objects
}
