% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plotting.R
\name{print.istat_plot_ready}
\alias{print.istat_plot_ready}
\title{Print Method for istat_plot_ready}
\usage{
\method{print}{istat_plot_ready}(x, ...)
}
\arguments{
\item{x}{An istat_plot_ready object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisible x
}
\description{
Print Method for istat_plot_ready
}
