% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tracking.R
\name{save_dataset_timestamps}
\alias{save_dataset_timestamps}
\title{Save Dataset Timestamps}
\usage{
save_dataset_timestamps(timestamps, timestamp_file = NULL)
}
\arguments{
\item{timestamps}{List containing timestamp registry data}

\item{timestamp_file}{Character string specifying path to timestamp registry file.
If NULL, uses default from configuration}
}
\value{
Invisible NULL
}
\description{
Saves the timestamp registry to a JSON file using atomic write operation
(write to temporary file then rename).
}
\examples{
\dontrun{
# Save timestamps
timestamps <- get_dataset_timestamps()
timestamps$datasets[["534_50"]] <- list(
  last_download = Sys.time(),
  row_count = 1000
)
save_dataset_timestamps(timestamps)
}
}
