% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{search_dataflows}
\alias{search_dataflows}
\title{Search Dataflows by Keywords}
\usage{
search_dataflows(
  keywords,
  fields = c("name", "description"),
  ignore_case = TRUE
)
}
\arguments{
\item{keywords}{Character vector of search terms}

\item{fields}{Character vector of fields to search in ("name", "description", "id")}

\item{ignore_case}{Logical indicating case-insensitive search}
}
\value{
A filtered data.table of matching dataflows
}
\description{
Searches available dataflows using keywords in Italian and English.
}
\examples{
\dontrun{
# Search for employment-related datasets
employment_datasets <- search_dataflows(c("employment", "lavoro", "occupazione"))

# Search unemployment datasets
unemployment_datasets <- search_dataflows(c("unemployment", "disoccupazione"))
}
}
