% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{validate_istat_data}
\alias{validate_istat_data}
\title{Validate ISTAT Data}
\usage{
validate_istat_data(data, required_cols = c("ObsDimension", "ObsValue"))
}
\arguments{
\item{data}{A data.table to validate}

\item{required_cols}{Character vector of required column names}
}
\value{
Logical indicating if data is valid
}
\description{
Validates the structure and content of ISTAT data.
}
\examples{
\dontrun{
is_valid <- validate_istat_data(my_data, c("tempo", "valore"))
}
}
